+++ proto/mount_options.go
@@ -283,7 +283,7 @@ type MountOptions struct {
 	NearRead             bool
 	EnablePosixACL       bool
 	EnableTransaction    string
-	TxTimeout            uint32
+	TxTimeout            int64
 	VolType              int
 	EbsEndpoint          string
 	EbsServicePath       string
